Add redux :(
Passing data between the DOM and Pixi environment is a bit tricky
because we need a separate provider for each but with the same store. I
didn't find a neat way of doing it with the context API, so I used Redux
instead, and then made a useData wrapper to make the usage the same as
context.